requirements.txt
torch==2.4.1
torchvision==0.19.1
pillow==10.4.0
pandas==2.2.2
numpy==1.26.4
nibabel==5.2.1
SimpleITK==2.4.0
tqdm==4.66.5
TotalSegmentator==2.4.0
nnunetv2==2.5.1